Break down `ItemPicker` styled file
